+++ test/config.js
@@ -41,6 +41,26 @@ const consumerConfig = Object.assign({}, config, {
   }
 });
 
+const jsProducerConfig = {
+  kafkaHost: "localhost:9092",
+  clientName: "n-test-producer-js",
+  options: {}
+};
+
+const jsConsumerConfig = {
+  kafkaHost: "localhost:9092",
+  groupId: "n-test-group-js",
+  options: {
+    autoCommit: true
+  }
+};
+
 const topic = "n-test-topic";
 
-module.exports = {producerConfig, consumerConfig, topic};
+module.exports = {
+  producerConfig, 
+  consumerConfig, 
+  jsProducerConfig, 
+  jsConsumerConfig, 
+  topic
+};